base/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
# cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.0
django-etc==1.4.0
django-hitcount==1.3.5
django-redirects==0.4.0
django-taggit==3.1.0
django-treebeard==4.6.1
docopt==0.6.2
idna==3.4
mutagen==1.46.0
nanoid==2.0.0
numerize==0.12
oauthlib==3.2.2
Pillow==9.4.0
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.6.0
python-slugify==8.0.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
urllib3==1.26.15
validate-email==1.3
yarg==0.1.9
